+++ x/evidence/genesis.go
@@ -2,6 +2,7 @@ package evidence
 
 import (
 	"context"
+	"errors"
 	"fmt"
 
 	"cosmossdk.io/x/evidence/exported"
@@ -13,28 +14,29 @@ import (
 
 // InitGenesis initializes the evidence module's state from a provided genesis
 // state.
-func InitGenesis(ctx context.Context, k keeper.Keeper, gs *types.GenesisState) {
+func InitGenesis(ctx context.Context, k keeper.Keeper, gs *types.GenesisState) error {
 	if err := gs.Validate(); err != nil {
-		panic(fmt.Sprintf("failed to validate %s genesis state: %s", types.ModuleName, err))
+		return fmt.Errorf("failed to validate %s genesis state: %s", types.ModuleName, err)
 	}
 
 	for _, e := range gs.Evidence {
 		evi, ok := e.GetCachedValue().(exported.Evidence)
 		if !ok {
-			panic("expected evidence")
+			return errors.New("expected evidence")
 		}
 		if _, err := k.Evidences.Get(ctx, evi.Hash()); err == nil {
-			panic(fmt.Sprintf("evidence with hash %s already exists", evi.Hash()))
+			return fmt.Errorf("evidence with hash %s already exists", evi.Hash())
 		}
 
 		if err := k.Evidences.Set(ctx, evi.Hash(), evi); err != nil {
-			panic(err)
+			return err
 		}
 	}
+	return nil
 }
 
 // ExportGenesis returns the evidence module's exported genesis.
-func ExportGenesis(ctx context.Context, k keeper.Keeper) *types.GenesisState {
+func ExportGenesis(ctx context.Context, k keeper.Keeper) (*types.GenesisState, error) {
 	gs := new(types.GenesisState)
 	err := k.Evidences.Walk(ctx, nil, func(_ []byte, value exported.Evidence) (stop bool, err error) {
 		anyEvi, err := codectypes.NewAnyWithValue(value)
@@ -45,7 +47,7 @@ func ExportGenesis(ctx context.Context, k keeper.Keeper) *types.GenesisState {
 		return false, nil
 	})
 	if err != nil {
-		panic(err)
+		return nil, err
 	}
-	return gs
+	return gs, nil
 }